move source_url_pattern to Utils
